moved pubky host to the shared module
