+++ pubky-homeserver/src/admin/routes/mod.rs
@@ -2,3 +2,4 @@ pub(crate) mod delete_entry;
 pub(crate) mod disable_users;
 pub(crate) mod generate_signup_token;
 pub(crate) mod root;
+pub(crate) mod nginx_auth_request;
