+++ pubky-homeserver/src/core/layers/mod.rs
@@ -1,3 +1,2 @@
 pub mod authz;
-pub mod pubky_host;
 pub mod trace;
