+++ pubky-homeserver/src/core/routes/tenants/session.rs
@@ -1,13 +1,12 @@
 use axum::{extract::State, http::StatusCode, response::IntoResponse};
 use tower_cookies::Cookies;
 
-use crate::core::{
+use crate::{core::{
     err_if_user_is_invalid::err_if_user_is_invalid,
     error::{Error, Result},
-    extractors::PubkyHost,
     layers::authz::session_secret_from_cookies,
     AppState,
-};
+}, shared::PubkyHost};
 
 pub async fn session(
     State(state): State<AppState>,
