+++ pubky-homeserver/src/core/routes/tenants/write.rs
@@ -8,12 +8,11 @@ use axum::{
 };
 use futures_util::stream::StreamExt;
 
-use crate::core::{
+use crate::{core::{
     err_if_user_is_invalid::err_if_user_is_invalid,
     error::{Error, Result},
-    extractors::PubkyHost,
     AppState,
-};
+}, shared::PubkyHost};
 
 /// Fail with 507 if `(current + incoming − existing) > quota`.
 fn enforce_user_disk_quota(
