+++ pubky-homeserver/src/shared/mod.rs
@@ -1,5 +1,7 @@
 mod http_error;
 mod pubkey_path_validator;
+mod pubky_host;
 
 pub(crate) use http_error::{HttpError, HttpResult};
 pub(crate) use pubkey_path_validator::Z32Pubkey;
+pub(crate) use pubky_host::{PubkyHost, PubkyHostLayer};
