+++ pubky-homeserver/src/shared/pubky_host.rs
@@ -1,12 +1,20 @@
-use crate::core::error::Result;
-use crate::core::extractors::PubkyHost;
 use axum::{body::Body, http::Request};
 use futures_util::future::BoxFuture;
 use pkarr::PublicKey;
 use std::{convert::Infallible, task::Poll};
 use tower::{Layer, Service};
+use std::fmt::Display;
+
+use axum::{
+    extract::FromRequestParts,
+    http::{request::Parts, StatusCode},
+    response::{IntoResponse, Response},
+};
+
 
 /// A Tower Layer to extract and inject the PubkyHost into request extensions.
+/// This is added to the router so the host is extracted automatically for each request.
+/// You then can use `PubkyHost` extractor to get the host from the request.
 #[derive(Debug, Clone)]
 pub struct PubkyHostLayer;
 
@@ -80,4 +88,44 @@ fn extract_pubky(req: &Request<Body>) -> Option<PublicKey> {
         });
     }
     pubky
+}
+
+
+
+
+/// Extractor for the PubkyHost.
+#[derive(Debug, Clone)]
+pub struct PubkyHost(pub(crate) PublicKey);
+
+impl PubkyHost {
+    pub fn public_key(&self) -> &PublicKey {
+        &self.0
+    }
+}
+
+impl Display for PubkyHost {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.0)
+    }
+}
+
+impl<S> FromRequestParts<S> for PubkyHost
+where
+    S: Sync + Send,
+{
+    type Rejection = Response;
+
+    async fn from_request_parts(parts: &mut Parts, _state: &S) -> Result<Self, Self::Rejection> {
+        let pubky_host = parts
+            .extensions
+            .get::<PubkyHost>()
+            .cloned()
+            .ok_or((
+                StatusCode::INTERNAL_SERVER_ERROR,
+                "Can't extract PubkyHost. Is `PubkyHostLayer` enabled?",
+            ))
+            .map_err(|e| e.into_response())?;
+
+        Ok(pubky_host)
+    }
 }
